Pre-custom msbuild
