+++ WORKSPACE
@@ -60,8 +60,8 @@ dotnet_register_toolchains(
         "windows_amd64": "abcd034b230365d9454459e271e118a851969d82516b1529ee0bfea07f7aae52",
         "linux_amd64": "3687b2a150cd5fef6d60a4693b4166994f32499c507cd04f346b6dda38ecdc46",
     },
-    version = "3.1.100",
-    #    version = "host",
+    #    version = "3.1.100",
+    version = "host",
 )
 
 load("//deps:nuget.bzl", "nuget_deps")
