+++ dotnet/private/msbuild/xml.bzl
@@ -101,7 +101,7 @@ def prepare_project_file(
         "MSBuildProjectExtensionsPath": THIS_DIR + intermediate_path,
         # we could just set ProjectAssetsFile here, but we're setting the other properties in case they have other impacts
         "OutputPath": THIS_DIR + paths.dirname(intermediate_path),
-        "ImportDirectoryBuildProps": "false",
+        #"ImportDirectoryBuildProps": "false",
         "UseSharedCompilation": "false",
     })
