+++ tests/examples/Transitive/TransitiveTransitiveLib/BUILD
@@ -1,3 +1,4 @@
+load("//tests/tools/build_test:build_test.bzl", "build_test")
 load("@my_rules_dotnet//dotnet:defs.bzl", "dotnet_library")
 
 dotnet_library(
@@ -5,4 +6,18 @@ dotnet_library(
     srcs = ["TransitiveTransitiveLibClass.cs"],
     target_framework = "netstandard2.1",
     visibility = ["//tests/examples/Transitive/TransitiveLib:__pkg__"],
+)
+
+build_test(
+    name = "TransitiveTransitiveLib_test",
+    expected_files = {
+        "": [
+            "TransitiveTransitiveLib.csproj.cache",
+        ],
+        "netstandard2.1": [
+            "TransitiveTransitiveLib.pdb",
+            "TransitiveTransitiveLib.dll",
+        ],
+    },
+    target = "TransitiveTransitiveLib",
 )
